Use 8, 16, 32, or 64 bits for integer union types based on number of cases.
Previously they were all 64 bits. This means that bools were 64 bits just to hold 0 or 1. Now they use 8 bits.

+++ flowgraph/interp/obj.go
@@ -525,13 +525,12 @@ func bin(o, p Number, op flowgraph.OpKind) Number {
 	}
 }
 
-// TODO: Bool may not always be Int64.
 // Instead, we should probably make bool always be int8 on all archs.
 func Bool(b bool) Number {
 	if b {
-		return Int64(1)
+		return Uint8(1)
 	}
-	return Int64(0)
+	return Uint8(0)
 }
 
 func convert(obj Val, typ flowgraph.Type) Val {
